mss_evaluation/mss-oracle-experiments/requirements-cupy.txt
scipy
numpy
cupy-cuda115
torch==1.9
git+git://github.com/sevagh/sigsep-mus-eval@feat/cupy-accel#egg=museval
git+git://github.com/sevagh/nsgt@main#egg=nsgt
librosa
pandas
seaborn
line_profiler
memory_profiler
tqdm
musdb
